# apps/aurora-core/build/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "aurora-core",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"servers": [
		{
			"url": "/api"
		}
	],
	"components": {
...
		"securitySchemes": {
			"local": {
				"type": "apiKey",
				"name": "connect.sid",
				"in": "cookie"
			},
			"integration": {
				"type": "apiKey",
				"name": "x-api-key",
				"in": "header"
			}
		},
		"schemas": {
...
			"Partial_Record_SocketioNamespaces.string__": {
				"properties": {
					"/": {
						"type": "string"
					},
					"/audio": {
						"type": "string"
					},
					"/screen": {
						"type": "string"
					},
					"/lights": {
						"type": "string"
					},
					"/public": {
						"type": "string"
					},
					"/backoffice": {
						"type": "string"
					},
					"/backoffice-beat": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
			"LightsControllerResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"socketIds": {
						"$ref": "#/components/schemas/Partial_Record_SocketioNamespaces.string__",
						"description": "ID of the socket connection if present.\nRequired to send events specifically and only to this entity.\nMapping from namespace to ID, as a websocket has a different ID\nfor each namespace its in."
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LightsControllerCreateParams": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LightsSwitchResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"dmxChannel": {
						"type": "number",
						"format": "double",
						"description": "DMX channel this lights switch is on"
					},
					"onValue": {
						"type": "number",
						"format": "double",
						"description": "DMX value to send to the channel to turn on the switch"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"name",
					"dmxChannel",
					"onValue"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ShutterChannelValuesResponse": {
				"properties": {
					"open": {
						"type": "number",
						"format": "double"
					},
					"strobe": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ParResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"nrChannels": {
						"type": "number",
						"format": "double"
					},
					"canReset": {
						"type": "boolean"
					},
					"resetChannel": {
						"type": "number",
						"format": "double"
					},
					"resetChannelValue": {
						"type": "number",
						"format": "double"
					},
					"shutterChannelValues": {
						"$ref": "#/components/schemas/ShutterChannelValuesResponse"
					},
					"masterDimChannel": {
						"type": "number",
						"format": "double"
					},
					"shutterChannel": {
						"type": "number",
						"format": "double"
					},
					"redChannel": {
						"type": "number",
						"format": "double"
					},
					"blueChannel": {
						"type": "number",
						"format": "double"
					},
					"greenChannel": {
						"type": "number",
						"format": "double"
					},
					"coldWhiteChannel": {
						"type": "number",
						"format": "double"
					},
					"warmWhiteChannel": {
						"type": "number",
						"format": "double"
					},
					"amberChannel": {
						"type": "number",
						"format": "double"
					},
					"uvChannel": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"name",
					"nrChannels",
					"canReset",
					"shutterChannelValues",
					"redChannel",
					"blueChannel",
					"greenChannel"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"MovingHeadRgbResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"nrChannels": {
						"type": "number",
						"format": "double"
					},
					"canReset": {
						"type": "boolean"
					},
					"resetChannel": {
						"type": "number",
						"format": "double"
					},
					"resetChannelValue": {
						"type": "number",
						"format": "double"
					},
					"shutterChannelValues": {
						"$ref": "#/components/schemas/ShutterChannelValuesResponse"
					},
					"panChannel": {
						"type": "number",
						"format": "double"
					},
					"finePanChannel": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"tiltChannel": {
						"type": "number",
						"format": "double"
					},
					"fineTiltChannel": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"movingSpeedChannel": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"masterDimChannel": {
						"type": "number",
						"format": "double"
					},
					"shutterChannel": {
						"type": "number",
						"format": "double"
					},
					"redChannel": {
						"type": "number",
						"format": "double"
					},
					"blueChannel": {
						"type": "number",
						"format": "double"
					},
					"greenChannel": {
						"type": "number",
						"format": "double"
					},
					"coldWhiteChannel": {
						"type": "number",
						"format": "double"
					},
					"warmWhiteChannel": {
						"type": "number",
						"format": "double"
					},
					"amberChannel": {
						"type": "number",
						"format": "double"
					},
					"uvChannel": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"name",
					"nrChannels",
					"canReset",
					"shutterChannelValues",
					"panChannel",
					"tiltChannel",
					"redChannel",
					"blueChannel",
					"greenChannel"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"MovingHeadWheelResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"nrChannels": {
						"type": "number",
						"format": "double"
					},
					"canReset": {
						"type": "boolean"
					},
					"resetChannel": {
						"type": "number",
						"format": "double"
					},
					"resetChannelValue": {
						"type": "number",
						"format": "double"
					},
					"shutterChannelValues": {
						"$ref": "#/components/schemas/ShutterChannelValuesResponse"
					},
					"panChannel": {
						"type": "number",
						"format": "double"
					},
					"finePanChannel": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"tiltChannel": {
						"type": "number",
						"format": "double"
					},
					"fineTiltChannel": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"movingSpeedChannel": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"masterDimChannel": {
						"type": "number",
						"format": "double"
					},
					"shutterChannel": {
						"type": "number",
						"format": "double"
					},
					"wheelColorChannel": {
						"type": "number",
						"format": "double"
					},
					"wheelColorChannelValues": {
						"items": {
							"$ref": "#/components/schemas/MovingHeadWheelColorChannelValueResponse"
						},
						"type": "array"
					},
					"wheelGoboChannel": {
						"type": "number",
						"format": "double"
					},
					"gobos": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"wheelGoboRotateChannel": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"goboRotates": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"name",
					"nrChannels",
					"canReset",
					"shutterChannelValues",
					"panChannel",
					"tiltChannel",
					"masterDimChannel",
					"wheelColorChannel",
					"wheelColorChannelValues",
					"wheelGoboChannel",
					"gobos",
					"wheelGoboRotateChannel",
					"goboRotates"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"LightsGroupResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"controller": {
						"$ref": "#/components/schemas/LightsControllerResponse"
					},
					"gridSizeX": {
						"type": "number",
						"format": "double"
					},
					"gridSizeY": {
						"type": "number",
						"format": "double"
					},
					"pars": {
						"items": {
							"$ref": "#/components/schemas/FixtureInGroupResponse_ParResponse_"
						},
						"type": "array"
					},
					"movingHeadRgbs": {
						"items": {
							"$ref": "#/components/schemas/FixtureInGroupResponse_MovingHeadRgbResponse_"
						},
						"type": "array"
					},
					"movingHeadWheels": {
						"items": {
							"$ref": "#/components/schemas/FixtureInGroupResponse_MovingHeadWheelResponse_"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"name",
					"controller",
					"gridSizeX",
					"gridSizeY",
					"pars",
					"movingHeadRgbs",
					"movingHeadWheels"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"LightsGroupCreateParams": {
				"properties": {
					"name": {
						"type": "string"
					},
					"defaultHandler": {
						"type": "string"
					},
					"groupInMiddle": {
						"type": "boolean",
						"description": "Whether the group of lights is situated in the middle of the room, or at the side.\nHandlers can use this to determine which effects to use."
					},
					"gridSizeX": {
						"type": "number",
						"format": "float",
						"description": "Size (width) of the X axis where all the fixtures are positioned.\nAll fixtures should have their positionX be in range [0, gridSizeX).",
						"minimum": 0
					},
					"gridSizeY": {
						"type": "number",
						"format": "float",
						"description": "Size (width) of the Y axis where all the fixtures are positioned.\n0 if the lights are positioned in a line (and not in a grid)",
						"minimum": 0
					},
					"pars": {
						"items": {
							"$ref": "#/components/schemas/LightsInGroup"
						},
						"type": "array"
					},
					"movingHeadRgbs": {
						"items": {
							"$ref": "#/components/schemas/LightsInGroup"
						},
						"type": "array"
					},
					"movingHeadWheels": {
						"items": {
							"$ref": "#/components/schemas/LightsInGroup"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"defaultHandler",
					"groupInMiddle",
					"gridSizeX",
					"pars",
					"movingHeadRgbs",
					"movingHeadWheels"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LightsSwitchCreateParams": {
				"properties": {
					"name": {
						"type": "string"
					},
					"dmxChannel": {
						"type": "number",
						"format": "double",
						"description": "DMX channel this lights switch is on"
					},
					"onValue": {
						"type": "number",
						"format": "double",
						"description": "DMX value to send to the channel to turn on the switch"
					}
				},
				"required": [
					"name",
					"dmxChannel",
					"onValue"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"LightsParCreateParams": {
				"properties": {
					"name": {
						"type": "string"
					},
					"nrChannels": {
						"type": "number",
						"format": "double"
					},
					"shutterOptionValues": {
						"$ref": "#/components/schemas/ShutterOptionValues"
					},
					"masterDimChannel": {
						"type": "number",
						"format": "double"
					},
					"shutterChannel": {
						"type": "number",
						"format": "double"
					},
					"colorRedChannel": {
						"type": "number",
						"format": "double"
					},
					"colorGreenChannel": {
						"type": "number",
						"format": "double"
					},
					"colorBlueChannel": {
						"type": "number",
						"format": "double"
					},
					"colorColdWhiteChannel": {
						"type": "number",
						"format": "double"
					},
					"colorWarmWhiteChannel": {
						"type": "number",
						"format": "double"
					},
					"colorAmberChannel": {
						"type": "number",
						"format": "double"
					},
					"colorUvChannel": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"nrChannels",
					"colorRedChannel",
					"colorGreenChannel",
					"colorBlueChannel"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LightsMovingHeadRgbCreateParams": {
				"properties": {
					"name": {
						"type": "string"
					},
					"nrChannels": {
						"type": "number",
						"format": "double"
					},
					"shutterOptionValues": {
						"$ref": "#/components/schemas/ShutterOptionValues"
					},
					"panChannel": {
						"type": "number",
						"format": "double"
					},
					"finePanChannel": {
						"type": "number",
						"format": "double"
					},
					"tiltChannel": {
						"type": "number",
						"format": "double"
					},
					"fineTiltChannel": {
						"type": "number",
						"format": "double"
					},
					"movingSpeedChannel": {
						"type": "number",
						"format": "double"
					},
					"basePanValue": {
						"type": "number",
						"format": "double"
					},
					"panMirrored": {
						"type": "boolean"
					},
					"tiltMirrored": {
						"type": "boolean"
					},
					"masterDimChannel": {
						"type": "number",
						"format": "double"
					},
					"shutterChannel": {
						"type": "number",
						"format": "double"
					},
					"colorRedChannel": {
						"type": "number",
						"format": "double"
					},
					"colorGreenChannel": {
						"type": "number",
						"format": "double"
					},
					"colorBlueChannel": {
						"type": "number",
						"format": "double"
					},
					"colorColdWhiteChannel": {
						"type": "number",
						"format": "double"
					},
					"colorWarmWhiteChannel": {
						"type": "number",
						"format": "double"
					},
					"colorAmberChannel": {
						"type": "number",
						"format": "double"
					},
					"colorUvChannel": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"nrChannels",
					"panChannel",
					"tiltChannel",
					"colorRedChannel",
					"colorGreenChannel",
					"colorBlueChannel"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"LightsColorResponse": {
				"properties": {
					"color": {
						"$ref": "#/components/schemas/RgbColor"
					},
					"spec": {
						"$ref": "#/components/schemas/RgbColorSpecification"
					}
				},
				"required": [
					"color",
					"spec"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		}
	},
	"paths": {
...
		"/lights/controller": {
			"get": {
				"operationId": "GetLightsControllers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/LightsControllerResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Lights"
				],
				"security": [
					{
						"local": [
							"admin",
							"board",
							"avico",
							"bac",
							"key-holder"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateLightsController",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LightsControllerResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Lights"
				],
				"security": [
					{
						"local": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LightsControllerCreateParams"
							}
						}
					}
				}
			}
		},
		"/lights/controller/{id}": {
			"get": {
				"operationId": "GetSingleLightsController",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LightsControllerResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Lights"
				],
				"security": [
					{
						"local": [
							"admin",
							"board",
							"avico",
							"bac",
							"key-holder"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/lights/switch": {
			"get": {
				"operationId": "GetAllLightsSwitches",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/LightsSwitchResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Lights"
				],
				"security": [
					{
						"local": [
							"admin",
							"board",
							"avico",
							"bac",
							"key-holder"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "enabled",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/lights/group": {
			"get": {
				"operationId": "GetLightsGroups",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/LightsGroupResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Lights"
				],
				"security": [
					{
						"local": [
							"admin",
							"board",
							"avico",
							"bac",
							"key-holder"
						]
					}
				],
				"parameters": []
			}
		},
		"/lights/group/{id}": {
			"get": {
				"operationId": "GetSingleLightsGroup",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LightsGroupResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Lights"
				],
				"security": [
					{
						"local": [
							"admin",
							"board",
							"avico",
							"bac",
							"key-holder"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/lights/controller/{id}/groups": {
			"get": {
				"operationId": "GetControllerLightsGroups",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/LightsGroupResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Lights"
				],
				"security": [
					{
						"local": [
							"lights-subscriber"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/lights/controller/{id}/group": {
			"post": {
				"operationId": "CreateLightsGroup",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LightsGroupResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Lights"
				],
				"security": [
					{
						"local": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LightsGroupCreateParams"
							}
						}
					}
				}
			}
		},
		"/lights/controller/{id}/switches": {
			"get": {
				"operationId": "GetControllerLightsSwitches",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/LightsSwitchResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Lights"
				],
				"security": [
					{
						"local": [
							"lights-subscriber"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateLightsSwitch",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LightsSwitchResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Lights"
				],
				"security": [
					{
						"local": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LightsSwitchCreateParams"
							}
						}
					}
				}
			}
		},
		"/lights/fixture/par": {
			"get": {
				"operationId": "GetAllLightsPars",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ParResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Lights"
				],
				"security": [
					{
						"local": [
							"admin"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateLightsPar",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ParResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Lights"
				],
				"security": [
					{
						"local": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LightsParCreateParams"
							}
						}
					}
				}
			}
		},
		"/lights/fixture/moving-head/rgb": {
			"get": {
				"operationId": "GetAllLightsMovingHeadsRgb",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MovingHeadRgbResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Lights"
				],
				"security": [
					{
						"local": [
							"admin"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateLightsMovingHeadRgb",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MovingHeadRgbResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Lights"
				],
				"security": [
					{
						"local": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LightsMovingHeadRgbCreateParams"
							}
						}
					}
				}
			}
		},
		"/lights/fixture/moving-head/wheel": {
			"get": {
				"operationId": "GetAllLightsMovingHeadsWheel",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MovingHeadWheelResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Lights"
				],
				"security": [
					{
						"local": [
							"admin"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateLightsMovingHeadWheel",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MovingHeadWheelResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Lights"
				],
				"security": [
					{
						"local": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LightsMovingHeadWheelCreateParams"
							}
						}
					}
				}
			}
		},
		"/lights/effects": {
			"get": {
				"operationId": "GetAllLightsEffects",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"type": "string"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Lights"
				],
				"security": [
					{
						"local": [
							"admin",
							"board",
							"avico",
							"bac",
							"key-holder"
						]
					}
				],
				"parameters": []
			}
		},
		"/lights/colors": {
			"get": {
				"operationId": "GetAllLightsColors",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/LightsColorResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Lights"
				],
				"security": [
					{
						"local": [
							"admin",
							"board",
							"avico",
							"bac",
							"key-holder",
							"audio-subscriber",
							"screen-subscriber",
							"lights-subscriber",
							"integration-user"
						]
					},
					{
						"integration": [
							"getAllLightsColors"
						]
					}
				],
				"parameters": []
			}
		},
...
	}
}